pyproject.toml
[project]
name = "ai-hallucination-agent"
version = "0.1.0"
description = "AI Hallucination Detection and Correction Agent"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "uvicorn",
    "requests",
    "python-dotenv",
    "openai",
    "openenv-core>=0.2.0"
]

[project.scripts]
server = "server.app:main"

[tool.openenv]
name = "AI Hallucination Detection Agent"
entrypoint = "server:app"
port = 7860